requirements.txt
# Core dependencies
google-cloud-aiplatform>=1.50.0
vertexai>=1.50.0
google-auth>=2.20.0
google-oauth2-tool>=0.0.3

# Image processing
Pillow>=10.0.0
pdf2image>=3.1.0

# Web requests
requests>=2.31.0

# Environment management
python-dotenv>=1.0.0

# Database (if needed for future features)
SQLAlchemy>=2.0.0

# Optional: For better PDF processing on Windows
# Note: Also requires poppler-utils installation
# Windows: choco install poppler
# Or download from: https://github.com/oschwartz10612/poppler-windows/releases/